requirements.txt
praw 
pandas 
python-dotenv
numpy
nltk
scikit-learn
joblib
requests
beautifulsoup4